Modern NBTify + SNBT
This is the modern take I think that's better for NBT editing over JSON editing. It's a much nicer experience that's closer to how the data is actually saved, and it's nicer to read I think too. The name of the project doesn't really line up with this anymore, but that's alright hehe.

An end of an era :')
I think he's strong for making the choice he needs too, I'm glad he's moving on to new adventures! It's always great to do new things and start new projects and make new goals.
https://www.youtube.com/watch?v=u1_9N3zxSbM