Named Imports (vs Namespaces)
Also, some modern formatting from how I've been doing things for a while now (the last year or so).

Made the script installable as a global shell command! I think that was the initial goal for this, I just didn't know how to do it yet. I want to streamline this so you can use it as a normal CLI command, rather than an interface that you 'go through menus' in a way, since that's currently how it works still.